refactor: extract SQL cache diagnostics into their own module
sql_caching.py had grown a log-redaction subsystem alongside the cache
flow. Move the redaction, parsing and formatting helpers to
sql_cache_diagnostics.py and colocate their tests, leaving sql_caching.py
as the read/write flow it started as.

Names crossing the new module boundary drop their leading underscore.
The repeated redact-and-bound idiom becomes _redacted_snippet, which
keeps the field-length bound private to the new module.
