+++ tests/unit/test_sql_caching.py
@@ -12,116 +12,31 @@ from parameterized import parameterized
 from pyarrow import ArrowInvalid
 
 from deepnote_toolkit.sql.sql_caching import (
-    _URL_QUERY_PATTERN,
     SqlCacheUpload,
-    _describe_exception,
-    _describe_presigned_url,
-    _describe_s3_error,
     _generate_cache_key,
-    _redact_sensitive,
     _request_cache_info_from_webapp,
-    _safe_url_path,
     get_sql_cache,
     upload_sql_cache,
 )
 from deepnote_toolkit.sql.sql_utils import is_single_select_query
 
+from .helpers.sql_cache_fixtures import (
+    ACCESS_DENIED_EXPIRED_BODY,
+    AWS_HEADERS,
+    EXPIRED_TOKEN_BODY,
+    PRESIGNED_URL,
+    PROXY_ECHO_BODY,
+    SECRETS,
+    SIGNATURE_MISMATCH_BODY,
+    URLLIB3_ERROR_MESSAGE,
+    s3_response,
+)
+
 QUERY = "SELECT * FROM users"
-
-# Signing parameters given values that are easy to search log output for
-PRESIGNED_URL = (
-    "https://bucket.s3.eu-west-1.amazonaws.com/ws/int/key"
-    "?X-Amz-Algorithm=AWS4-HMAC-SHA256"
-    "&X-Amz-Credential=CREDVALUE%2F20260729%2Feu-west-1%2Fs3%2Faws4_request"
-    "&X-Amz-Date=20260729T120000Z&X-Amz-Expires=900"
-    "&X-Amz-Security-Token=TOKENVALUE&X-Amz-Signature=SIGVALUE"
-)
-
-# What requests raises when it cannot reach the object store. The URL urllib3
-# embeds is path-only - there is no scheme and no host in front of it.
-URLLIB3_ERROR_MESSAGE = (
-    "HTTPSConnectionPool(host='bucket.s3.eu-west-1.amazonaws.com', port=443): "
-    "Max retries exceeded with url: /ws/int/key"
-    "?X-Amz-Algorithm=AWS4-HMAC-SHA256"
-    "&X-Amz-Credential=CREDVALUE%2F20260729%2Feu-west-1%2Fs3%2Faws4_request"
-    "&X-Amz-Date=20260729T120000Z&X-Amz-Expires=900"
-    "&X-Amz-Security-Token=TOKENVALUE&X-Amz-Signature=SIGVALUE "
-    "(Caused by NameResolutionError('Failed to resolve host'))"
-)
-
-SECRETS = ("CREDVALUE", "TOKENVALUE", "SIGVALUE", "X-Amz-")
-
-AWS_HEADERS = {
-    "x-amz-request-id": "REQ123",
-    "x-amz-id-2": "HOSTID456",
-    "Date": "Wed, 29 Jul 2026 12:31:07 GMT",
-}
-
-ACCESS_DENIED_EXPIRED_BODY = (
-    b'<?xml version="1.0" encoding="UTF-8"?>\n'
-    b"<Error><Code>AccessDenied</Code><Message>Request has expired</Message>"
-    b"<X-Amz-Expires>900</X-Amz-Expires>"
-    b"<Expires>2026-07-29T12:15:00Z</Expires>"
-    b"<ServerTime>2026-07-29T12:31:07Z</ServerTime>"
-    b"<RequestId>REQ123</RequestId><HostId>HOSTID456</HostId></Error>"
-)
-
-# S3 rejecting credentials that died before the URL's own expiry
-EXPIRED_TOKEN_BODY = (
-    b'<?xml version="1.0" encoding="UTF-8"?>\n'
-    b"<Error><Code>ExpiredToken</Code>"
-    b"<Message>The provided token has expired.</Message>"
-    b"<RequestId>REQ123</RequestId><HostId>HOSTID456</HostId></Error>"
-)
-
-# S3 echoes the canonical request - and with it the signed query string - when
-# the signature does not match
-SIGNATURE_MISMATCH_BODY = (
-    b'<?xml version="1.0" encoding="UTF-8"?>\n'
-    b"<Error><Code>SignatureDoesNotMatch</Code>"
-    b"<Message>The request signature we calculated does not match the signature "
-    b"you provided. Check your key and signing method.</Message>"
-    b"<AWSAccessKeyId>CREDVALUE</AWSAccessKeyId>"
-    b"<StringToSign>AWS4-HMAC-SHA256\n20260729T120000Z\n</StringToSign>"
-    b"<CanonicalRequest>PUT\n/ws/int/key\n"
-    b"X-Amz-Credential=CREDVALUE&amp;X-Amz-Security-Token=TOKENVALUE"
-    b"&amp;X-Amz-Signature=SIGVALUE\nhost:bucket.s3.eu-west-1.amazonaws.com\n"
-    b"</CanonicalRequest>"
-    b"<RequestId>REQ123</RequestId><HostId>HOSTID456</HostId></Error>"
-)
-
-# A gateway that answered instead of S3 and echoed the request line back
-PROXY_ECHO_BODY = (
-    b"<html><head><title>502 Bad Gateway</title></head><body>\n"
-    b"<h1>502 Bad Gateway</h1>\n"
-    b"<p>Upstream failed for request: PUT "
-    b"https://bucket.s3.eu-west-1.amazonaws.com/ws/int/key"
-    b"?X-Amz-Algorithm=AWS4-HMAC-SHA256"
-    b"&X-Amz-Credential=CREDVALUE%2F20260729%2Feu-west-1%2Fs3%2Faws4_request"
-    b"&X-Amz-Date=20260729T120000Z&X-Amz-Expires=900"
-    b"&X-Amz-Security-Token=TOKENVALUE&X-Amz-Signature=SIGVALUE</p>\n"
-    b"</body></html>"
-)
 
 RESERVED_LOGRECORD_ATTRS = set(
     logging.LogRecord("", 0, "", 0, "", None, None).__dict__
 ) | {"message", "asctime"}
-
-
-def _s3_response(status_code, body=b"", headers=None):
-    """Build a stand-in for a requests.Response from the object store."""
-    response = mock.MagicMock(
-        status_code=status_code, content=body, headers=headers or {}
-    )
-    response.__enter__.return_value = response
-    # a real streamed body reads back empty once the block exits, silently
-    response.__exit__.side_effect = lambda *_: setattr(response, "content", b"")
-    # a chunked response yields one HTTP chunk per read however large a size is
-    # asked for
-    response.iter_content.side_effect = lambda size: iter(
-        [body[i : i + 20] for i in range(0, len(body), 20)]
-    )
-    return response
 
 
 def _upload(url=PRESIGNED_URL, issued_at=None):
@@ -162,7 +77,7 @@ def _collect_logged_extras():
 
     with patch("deepnote_toolkit.sql.sql_caching.logger") as mock_logger:
         with patch("deepnote_toolkit.sql.sql_caching.requests.put") as mock_put:
-            mock_put.return_value = _s3_response(
+            mock_put.return_value = s3_response(
                 403, ACCESS_DENIED_EXPIRED_BODY, AWS_HEADERS
             )
             upload_sql_cache(dataframe, _upload())
@@ -181,7 +96,7 @@ def _collect_logged_extras():
         ) as mock_cache_info:
             mock_cache_info.return_value = _cache_hit()
             with patch("deepnote_toolkit.sql.sql_caching.requests.get") as mock_get:
-                mock_get.return_value = _s3_response(
+                mock_get.return_value = s3_response(
                     403, EXPIRED_TOKEN_BODY, AWS_HEADERS
                 )
                 get_sql_cache(QUERY, {}, "123", "read", "dataframe")
@@ -206,7 +121,7 @@ def _collect_logged_extras():
                 "?sqlCacheKey=abc&sqlCacheMode=read"
             )
             mock_headers.return_value = {}
-            mock_get.return_value = _s3_response(503, b"upstream unavailable")
+            mock_get.return_value = s3_response(503, b"upstream unavailable")
             _request_cache_info_from_webapp("abc", "123", "read")
 
         return [call.kwargs["extra"] for call in mock_logger.error.call_args_list]
@@ -347,7 +262,7 @@ class TestGetSqlCache(unittest.TestCase):
 
         mock_is_single_select_query.return_value = True
         mock_request_cache_info_from_webapp.return_value = cache_info
-        mock_get.return_value = _s3_response(200, b"parquet-bytes")
+        mock_get.return_value = s3_response(200, b"parquet-bytes")
         mock_read_parquet.return_value = pd.DataFrame()
 
         result_df, upload = get_sql_cache(
@@ -394,7 +309,7 @@ class TestGetSqlCache(unittest.TestCase):
 
         mock_is_single_select_query.return_value = True
         mock_request_cache_info_from_webapp.return_value = cache_info
-        mock_get.return_value = _s3_response(200, b"pickle-bytes")
+        mock_get.return_value = s3_response(200, b"pickle-bytes")
         mock_read_parquet.side_effect = ArrowInvalid
         mock_read_pickle.return_value = pd.DataFrame()
 
@@ -446,7 +361,7 @@ class TestGetSqlCache(unittest.TestCase):
 
         mock_is_single_select_query.return_value = True
         mock_request_cache_info_from_webapp.return_value = cache_info
-        mock_get.return_value = _s3_response(200, b"parquet-bytes")
+        mock_get.return_value = s3_response(200, b"parquet-bytes")
         mock_read_parquet.side_effect = Exception("Failed to download from cache")
 
         result_df, upload = get_sql_cache(
@@ -545,7 +460,7 @@ class TestGetSqlCache(unittest.TestCase):
         mock_logger,
     ):
         mock_cache_info.return_value = _cache_hit("https://example.com/cache")
-        mock_get.return_value = _s3_response(200, b"not-a-dataframe")
+        mock_get.return_value = s3_response(200, b"not-a-dataframe")
         mock_read_parquet.side_effect = ArrowInvalid
         mock_read_pickle.side_effect = Exception("Error reading pickle")
 
@@ -564,7 +479,7 @@ class TestGetSqlCache(unittest.TestCase):
         self, mock_get, mock_cache_info, mock_logger
     ):
         mock_cache_info.return_value = _cache_hit()
-        mock_get.return_value = _s3_response(403, EXPIRED_TOKEN_BODY, AWS_HEADERS)
+        mock_get.return_value = s3_response(403, EXPIRED_TOKEN_BODY, AWS_HEADERS)
 
         result_df, upload = get_sql_cache(QUERY, {}, "123", "read", "dataframe")
 
@@ -597,7 +512,7 @@ class TestGetSqlCache(unittest.TestCase):
         self, _name, body, mock_get, mock_cache_info, mock_logger
     ):
         mock_cache_info.return_value = _cache_hit()
-        mock_get.return_value = _s3_response(403, body, AWS_HEADERS)
+        mock_get.return_value = s3_response(403, body, AWS_HEADERS)
 
         get_sql_cache(QUERY, {}, "123", "read", "dataframe")
 
@@ -617,7 +532,7 @@ class TestGetSqlCache(unittest.TestCase):
         dataframe.to_parquet(buffer)
 
         mock_cache_info.return_value = _cache_hit()
-        mock_get.return_value = _s3_response(200, buffer.getvalue())
+        mock_get.return_value = s3_response(200, buffer.getvalue())
 
         result_df, _ = get_sql_cache(QUERY, {}, "123", "read", "dataframe")
 
@@ -637,7 +552,7 @@ class TestGetSqlCache(unittest.TestCase):
         dataframe.to_pickle(buffer)
 
         mock_cache_info.return_value = _cache_hit()
-        mock_get.return_value = _s3_response(200, buffer.getvalue())
+        mock_get.return_value = s3_response(200, buffer.getvalue())
 
         result_df, _ = get_sql_cache(QUERY, {}, "123", "read", "dataframe")
 
@@ -657,7 +572,7 @@ class TestGetSqlCache(unittest.TestCase):
         dataframe.to_parquet(buffer)
 
         mock_cache_info.return_value = _cache_hit()
-        mock_get.return_value = _s3_response(200, buffer.getvalue())
+        mock_get.return_value = s3_response(200, buffer.getvalue())
 
         get_sql_cache(QUERY, {}, "123", "read", "dataframe")
 
@@ -703,7 +618,7 @@ class TestRequestCacheInfoFromWebapp(unittest.TestCase):
             "?sqlCacheKey=abc&sqlCacheMode=read"
         )
         mock_headers.return_value = {}
-        mock_get.return_value = _s3_response(503, b"upstream unavailable")
+        mock_get.return_value = s3_response(503, b"upstream unavailable")
 
         self.assertIsNone(_request_cache_info_from_webapp("abc", "123", "read"))
 
@@ -718,244 +633,6 @@ class TestRequestCacheInfoFromWebapp(unittest.TestCase):
             "/userpod-api/p1/integrations/123/sql-cache",
         )
         self.assertNotIn("sqlCacheKey", " ".join(_logged_strings(mock_logger)))
-
-
-class TestRedactSensitive(unittest.TestCase):
-    def test_strips_query_string_from_urlopen_style_message(self):
-        # the shape urllib3 actually produces: the URL is path-only
-        redacted = _redact_sensitive(URLLIB3_ERROR_MESSAGE)
-
-        self.assertIn("bucket.s3.eu-west-1.amazonaws.com", redacted)
-        self.assertIn("/ws/int/key?<redacted>", redacted)
-        for secret in SECRETS:
-            self.assertNotIn(secret, redacted)
-
-    def test_query_string_strip_alone_redacts_urllib3_message(self):
-        """The primary defence must hold without help from the parameter backstop."""
-        stripped = _URL_QUERY_PATTERN.sub(r"\1?<redacted>", URLLIB3_ERROR_MESSAGE)
-
-        for secret in SECRETS:
-            self.assertNotIn(secret, stripped)
-
-    def test_blanks_aws_params_outside_a_url(self):
-        redacted = _redact_sensitive(
-            "X-Amz-Credential=CREDVALUE&X-Amz-Security-Token=TOKENVALUE"
-            "&X-Amz-Signature=SIGVALUE"
-        )
-
-        self.assertEqual(
-            redacted,
-            "X-Amz-Credential=<redacted>&X-Amz-Security-Token=<redacted>"
-            "&X-Amz-Signature=<redacted>",
-        )
-
-    @parameterized.expand(
-        [
-            ("question_in_prose", "Is this ok? Yes it is."),
-            ("bare_question", "what? nothing"),
-            ("plain_sentence", "The provided token has expired."),
-        ]
-    )
-    def test_leaves_ordinary_text_unchanged(self, _, text):
-        self.assertEqual(_redact_sensitive(text), text)
-
-
-class TestDescribeException(unittest.TestCase):
-    def test_large_message_is_bounded_and_redacted_in_bounded_time(self):
-        """The message is cut before redacting, not after.
-
-        _URL_QUERY_PATTERN backtracks over every start position in a long run of
-        non-separator characters, so redacting an untruncated exception message is
-        quadratic - 64k characters took 8.7s, charged to the user's cell.
-        """
-        message = URLLIB3_ERROR_MESSAGE + "&padding=" + "A" * 64_000
-
-        started = time.monotonic()
-        described = _describe_exception(ValueError(message))
-        elapsed = time.monotonic() - started
-
-        self.assertEqual(described["error_type"], "ValueError")
-        self.assertLessEqual(len(described["error_message"]), 500)
-        # cutting the query string short is still safe: the pattern runs to the
-        # end of the string, so the remainder is stripped either way
-        self.assertIn("/ws/int/key?<redacted>", described["error_message"])
-        for secret in SECRETS:
-            self.assertNotIn(secret, described["error_message"])
-        self.assertLess(elapsed, 2.0)
-
-
-class TestDescribeS3Error(unittest.TestCase):
-    def test_extracts_code_and_message_from_xml(self):
-        diagnostics = _describe_s3_error(
-            _s3_response(403, ACCESS_DENIED_EXPIRED_BODY, AWS_HEADERS)
-        )
-
-        self.assertEqual(diagnostics["status_code"], 403)
-        self.assertEqual(diagnostics["s3_error_code"], "AccessDenied")
-        self.assertIn("Request has expired", diagnostics["s3_error_message"])
-        # AWS's own account of when the URL died and what time it was
-        self.assertEqual(diagnostics["s3_expires"], "2026-07-29T12:15:00Z")
-        self.assertEqual(diagnostics["s3_server_time"], "2026-07-29T12:31:07Z")
-
-    def test_extracts_expired_token_body(self):
-        diagnostics = _describe_s3_error(_s3_response(403, EXPIRED_TOKEN_BODY))
-
-        self.assertEqual(diagnostics["s3_error_code"], "ExpiredToken")
-        self.assertEqual(
-            diagnostics["s3_error_message"], "The provided token has expired."
-        )
-
-    def test_captures_aws_request_headers(self):
-        diagnostics = _describe_s3_error(
-            _s3_response(403, EXPIRED_TOKEN_BODY, AWS_HEADERS)
-        )
-
-        self.assertEqual(diagnostics["aws_request_id"], "REQ123")
-        self.assertEqual(diagnostics["aws_host_id"], "HOSTID456")
-        self.assertEqual(diagnostics["aws_date"], "Wed, 29 Jul 2026 12:31:07 GMT")
-
-    def test_signature_mismatch_body_surfaces_only_code_and_message(self):
-        """Pins the field allowlist: <CanonicalRequest> and friends are never read.
-
-        Redaction is not what protects this body - the elements carrying the signed
-        query string are simply not among the four that get extracted.
-        """
-        diagnostics = _describe_s3_error(
-            _s3_response(403, SIGNATURE_MISMATCH_BODY, AWS_HEADERS)
-        )
-
-        self.assertEqual(diagnostics["s3_error_code"], "SignatureDoesNotMatch")
-        self.assertNotIn("response_body_snippet", diagnostics)
-        for value in diagnostics.values():
-            for secret in SECRETS:
-                self.assertNotIn(secret, str(value))
-
-    def test_proxy_body_echoing_request_url_is_redacted(self):
-        """Pins redaction: the allowlist cannot help once the body is not S3's."""
-        diagnostics = _describe_s3_error(_s3_response(502, PROXY_ECHO_BODY))
-
-        self.assertIsNone(diagnostics["s3_error_code"])
-        snippet = diagnostics["response_body_snippet"]
-        # the request line survives, the credentials on it do not
-        self.assertIn("502 Bad Gateway", snippet)
-        self.assertIn("/ws/int/key?<redacted>", snippet)
-        for secret in SECRETS:
-            self.assertNotIn(secret, snippet)
-
-    def test_non_xml_body_yields_snippet_without_code(self):
-        diagnostics = _describe_s3_error(
-            _s3_response(502, b"<html><body>502 Bad Gateway</body></html>")
-        )
-
-        self.assertIsNone(diagnostics["s3_error_code"])
-        self.assertIsNone(diagnostics["aws_request_id"])
-        self.assertIn("502 Bad Gateway", diagnostics["response_body_snippet"])
-        self.assertLessEqual(len(diagnostics["response_body_snippet"]), 500)
-
-    def test_oversized_body_is_bounded(self):
-        body = (
-            b"<Error><Code>AccessDenied</Code><Message>"
-            + b"x" * 1000
-            + b"</Message>"
-            + b"y" * 10_000
-            + b"</Error>"
-        )
-
-        diagnostics = _describe_s3_error(_s3_response(403, body))
-
-        self.assertEqual(diagnostics["s3_error_code"], "AccessDenied")
-        for value in diagnostics.values():
-            if isinstance(value, str):
-                self.assertLessEqual(len(value), 500)
-
-    def test_body_prefix_is_streamed_rather_than_buffered(self):
-        """Reading .content downloads the whole body just to keep 4 KB of it."""
-        buffered = []
-        response = mock.MagicMock(status_code=403, headers={})
-        type(response).content = mock.PropertyMock(
-            side_effect=lambda: buffered.append("content")
-        )
-        response.iter_content.side_effect = lambda size: iter(
-            [ACCESS_DENIED_EXPIRED_BODY[:size]]
-        )
-
-        diagnostics = _describe_s3_error(response)
-
-        self.assertEqual(buffered, [])
-        self.assertEqual(diagnostics["s3_error_code"], "AccessDenied")
-
-
-class TestDescribePresignedUrl(unittest.TestCase):
-    def test_returns_path_and_expiry(self):
-        described = _describe_presigned_url(PRESIGNED_URL)
-
-        self.assertEqual(described["object_host"], "bucket.s3.eu-west-1.amazonaws.com")
-        self.assertEqual(described["object_path"], "/ws/int/key")
-        self.assertEqual(described["url_expires_in"], 900)
-
-    def test_missing_expires_yields_none(self):
-        described = _describe_presigned_url("https://example.com/x")
-
-        self.assertEqual(described["object_path"], "/x")
-        self.assertIsNone(described["url_expires_in"])
-
-    def test_non_numeric_expires_yields_none(self):
-        described = _describe_presigned_url("https://example.com/x?X-Amz-Expires=abc")
-
-        self.assertIsNone(described["url_expires_in"])
-
-    @parameterized.expand(
-        [
-            ("unterminated_ipv6", "https://[::1"),
-            ("empty", ""),
-            ("not_a_url", "not a url"),
-            ("none", None),
-            ("bytes", b"/ws/int/key"),
-            ("dict", {"url": "https://example.com/x"}),
-        ]
-    )
-    def test_malformed_url_does_not_raise_or_leak(self, _, url):
-        described = _describe_presigned_url(url)
-
-        self.assertIsNone(described["url_expires_in"])
-        for value in described.values():
-            for secret in SECRETS:
-                self.assertNotIn(secret, str(value))
-        # a bytes value in either would silently discard the entire error report
-        json.dumps(described)
-        json.dumps(_safe_url_path(url))
-
-    @parameterized.expand(
-        [
-            (
-                "separator_encoded",
-                "%3FX-Amz-Credential=CREDVALUE&X-Amz-Security-Token=TOKENVALUE"
-                "&X-Amz-Signature=SIGVALUE",
-            ),
-            (
-                "separator_and_equals_encoded",
-                "%3FX-Amz-Credential%3DCREDVALUE&X-Amz-Security-Token%3DTOKENVALUE"
-                "&X-Amz-Signature%3DSIGVALUE",
-            ),
-            (
-                "whole_query_encoded",
-                "%3FX-Amz-Credential%3DCREDVALUE%26X-Amz-Security-Token%3DTOKENVALUE"
-                "%26X-Amz-Signature%3DSIGVALUE",
-            ),
-            ("separator_is_a_semicolon", ";X-Amz-Credential=CREDVALUE"),
-            ("no_separator_at_all", "X-Amz-Signature=SIGVALUE"),
-        ]
-    )
-    def test_over_encoded_url_does_not_leak_signing_params_via_path(self, _, suffix):
-        """urlsplit only splits on a literal '?', so the path carries the rest."""
-        described = _describe_presigned_url(
-            "https://bucket.s3.eu-west-1.amazonaws.com/ws/int/key" + suffix
-        )
-
-        # the parameter names may survive redaction, their values must not
-        for value in described.values():
-            for secret in ("CREDVALUE", "TOKENVALUE", "SIGVALUE"):
-                self.assertNotIn(secret, str(value))
 
 
 class TestUploadSqlCache(unittest.TestCase):
@@ -1040,7 +717,7 @@ class TestUploadSqlCache(unittest.TestCase):
     @patch("deepnote_toolkit.sql.sql_caching.logger")
     @patch("deepnote_toolkit.sql.sql_caching.requests.put")
     def test_http_error_logs_s3_diagnostics(self, mock_put, mock_logger):
-        mock_put.return_value = _s3_response(
+        mock_put.return_value = s3_response(
             403, ACCESS_DENIED_EXPIRED_BODY, AWS_HEADERS
         )
 
@@ -1074,7 +751,7 @@ class TestUploadSqlCache(unittest.TestCase):
     def test_http_error_logs_no_presigned_query_string(
         self, _name, body, mock_put, mock_logger
     ):
-        mock_put.return_value = _s3_response(403, body, AWS_HEADERS)
+        mock_put.return_value = s3_response(403, body, AWS_HEADERS)
 
         upload_sql_cache(pd.DataFrame({"a": [1, 2, 3]}), _upload())
 
@@ -1105,9 +782,9 @@ class TestUploadSqlCache(unittest.TestCase):
         df = pd.DataFrame({"a": [1, 2, 3]})
         other_url = PRESIGNED_URL.replace("/ws/int/key", "/other/int/key2")
 
-        mock_put.return_value = _s3_response(403, ACCESS_DENIED_EXPIRED_BODY)
+        mock_put.return_value = s3_response(403, ACCESS_DENIED_EXPIRED_BODY)
         upload_sql_cache(df, _upload())
-        mock_put.return_value = _s3_response(500, b"<Error><Code>Slow</Code></Error>")
+        mock_put.return_value = s3_response(500, b"<Error><Code>Slow</Code></Error>")
         upload_sql_cache(df, _upload(other_url))
         mock_put.side_effect = requests.exceptions.ConnectionError(
             URLLIB3_ERROR_MESSAGE
@@ -1138,7 +815,7 @@ class TestUploadSqlCache(unittest.TestCase):
     def test_seconds_since_url_issued_reflects_elapsed_time(
         self, mock_put, mock_logger
     ):
-        mock_put.return_value = _s3_response(403, ACCESS_DENIED_EXPIRED_BODY)
+        mock_put.return_value = s3_response(403, ACCESS_DENIED_EXPIRED_BODY)
 
         upload_sql_cache(
             pd.DataFrame({"a": [1, 2, 3]}),
@@ -1175,7 +852,7 @@ class TestUploadSqlCache(unittest.TestCase):
 
         def slow_put(*args, **kwargs):
             clock[0] += 1000.0
-            return _s3_response(500, b"<Error><Code>InternalError</Code></Error>")
+            return s3_response(500, b"<Error><Code>InternalError</Code></Error>")
 
         mock_put.side_effect = slow_put
 
@@ -1193,7 +870,7 @@ class TestUploadSqlCache(unittest.TestCase):
     @patch("deepnote_toolkit.sql.sql_caching.logger")
     @patch("deepnote_toolkit.sql.sql_caching.requests.put")
     def test_upload_request_bounds_connect_and_read_phases(self, mock_put, mock_logger):
-        mock_put.return_value = _s3_response(200)
+        mock_put.return_value = s3_response(200)
 
         upload_sql_cache(pd.DataFrame({"a": [1, 2, 3]}), _upload())
